refactor: fix conflict between local and remote in repository for search movies by country
