+++ app/src/main/java/com/berlin/aflami/MainActivity.kt
@@ -4,18 +4,11 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.navigationBarsPadding
-import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.berlin.aflami.navigation.AflamiNavGraph
-import com.berlin.aflami.screens.search.worldtour.WorldTourScreen
+import com.berlin.aflami.screens.search.country.SearchByCountryScreen
 import com.berlin.aflami.ui.theme.AflamiTheme
-import com.berlin.aflami.ui.theme.Theme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -34,7 +27,7 @@ class MainActivity : ComponentActivity() {
 //                        .statusBarsPadding()
 //                        .navigationBarsPadding()
 //                )
-                WorldTourScreen(
+                SearchByCountryScreen(
                     navController = navController
                 )
             }
