+++ app/src/main/java/com/berlin/aflami/navigation/routes/SearchByCountryRoute.kt
@@ -3,13 +3,13 @@ package com.berlin.aflami.navigation.routes
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.berlin.aflami.screens.search.worldtour.SearchByCountryScreen
-import com.example.navigation.Destination.WorldTourScreen
+import com.berlin.aflami.screens.search.country.SearchByCountryScreen
+import com.example.navigation.Destination.SearchByCountryScreen
 
 fun NavGraphBuilder.searchByCountryRoute(
     navController: NavController
 ) {
-    composable(route = WorldTourScreen.route) {
+    composable(route = SearchByCountryScreen.route) {
         SearchByCountryScreen(navController = navController)
     }
 }
