+++ presentation/navigation/src/main/java/com/example/navigation/Destination.kt
@@ -2,7 +2,7 @@ package com.example.navigation
 
 sealed class Destination(val route: String) {
     object SearchScreen : Destination("searchScreen")
-    object WorldTourScreen : Destination("worldTourScreen")
+    object SearchByCountryScreen : Destination("searchByCountryScreen")
     object SearchByActorNameScreen : Destination("searchByActorNameScreen")
 }
