+++ presentation/ui/src/main/kotlin/com/berlin/aflami/screens/search/country/composable/AnimatedCountriesList.kt
@@ -1,4 +1,4 @@
-package com.berlin.aflami.screens.search.worldtour.composable
+package com.berlin.aflami.screens.search.country.composable
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
@@ -11,7 +11,6 @@ import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Text
