+++ presentation/ui/src/main/kotlin/com/berlin/aflami/screens/search/screen/SearchScreen.kt
@@ -173,7 +173,7 @@ private fun SearchScreenContent(
 
                     SearchSuggestionHub(
                         Modifier.padding(horizontal = 16.dp),
-                        onWorldTourClick = { navController.navigate(Destination.WorldTourScreen.route) },
+                        onSearchByCountryClick = { navController.navigate(Destination.SearchByCountryScreen.route) },
                         onSearchByActorClick = { navController.navigate(Destination.SearchByActorNameScreen.route) })
 
                     NoDataSearch()
