+++ presentation/viewModel/src/test/kotlin/com/berlin/aflami/viewmodel/searchcountry/SearchByCountryViewModelTest.kt
@@ -21,7 +21,7 @@ import org.junit.Test
 import org.junit.jupiter.api.AfterEach
 import usecase.SearchByCountryUseCase
 
-class WorldTourViewModelTest {
+class SearchByCountryViewModelTest {
 
     private lateinit var viewModel: SearchByCountryViewModel
     private lateinit var searchByCountryUseCase: SearchByCountryUseCase
